+++ src/main/java/org/example/petwif/service/albumService/AlbumLikeServiceImpl.java
@@ -25,6 +25,8 @@ public class AlbumLikeServiceImpl implements AlbumLikeService{
     private final AlbumRepository albumRepository;
     private final MemberRepository memberRepository;
     private final AlbumLikeRepository albumLikeRepository;
+    private final AlbumCheckAccessService albumCheckAccessService;
+
 
 
     //좋아요 누르기
@@ -32,13 +34,15 @@ public class AlbumLikeServiceImpl implements AlbumLikeService{
     public void addAlbumLike(Long albumId, Long memberId){
         Album album = albumRepository.findById(albumId)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.ALBUM_NOT_FOUND));
+        albumCheckAccessService.checkAccess(album, memberId);
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(()-> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
+
         //좋아요 중복 방지 => 손봐야함
-        albumLikeRepository.findByAlbumAndMember(album, member)
-                .ifPresent(like -> {
-                    throw new IllegalStateException("이미 좋아요를 눌렀습니다.");
-                });
+        albumLikeRepository.findByAlbumAndMember(album, member).ifPresent(report -> {
+            throw new GeneralException(ErrorStatus.ALBUM_LIKE_EXIST);
+        });
+
 
         AlbumLike albumLike = AlbumLike.builder()
                 .album(album)
@@ -52,22 +56,26 @@ public class AlbumLikeServiceImpl implements AlbumLikeService{
     public void deleteAlbumLike(Long albumId, Long memberId){
         Album album = albumRepository.findById(albumId)
                 .orElseThrow(()-> new GeneralException(ErrorStatus.ALBUM_NOT_FOUND));
+        albumCheckAccessService.checkAccess(album, memberId);
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(()-> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
         AlbumLike albumLike = albumLikeRepository.findByAlbumAndMember(album, member)
-                .orElseThrow(() -> new IllegalStateException("좋아요가 없습니다."));
+                .orElseThrow(() -> new GeneralException(ErrorStatus.ALBUM_LIKE_NOT_FOUND));
 
        albumLikeRepository.delete(albumLike);
     }
 
     //좋아요 목록 보이기
-    public List<AlbumResponseDto.LikeResultDto> getAlbumLikes(Long albumId) {
+    public List<AlbumResponseDto.LikeResultDto> getAlbumLikes(Long albumId, Long memberId) {
         Album album = albumRepository.findById(albumId)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.ALBUM_NOT_FOUND));
 
+        albumCheckAccessService.checkAccess(album, memberId);
+
         List<AlbumLike> likes = albumLikeRepository.findByAlbum(album);
+        if(likes.isEmpty())throw new GeneralException(ErrorStatus.ALBUM_LIKE_PAGE_NOT_FOUND);
         return likes.stream()
-                .map(like -> new AlbumResponseDto.LikeResultDto(like.getMember().getId(), like.getMember().getName(), like.getMember().getProfile_url()))
+                .map(like -> new AlbumResponseDto.LikeResultDto(like.getMember().getId(), like.getMember().getNickname(), like.getMember().getProfile_url()))
                 .collect(Collectors.toList());
     }
